==> code/01_peotransfer.ipynb <==
import os
import glob
import rasterio
import fiona

import numpy as np
import pandas as pd
import geopandas as gp
import matplotlib.pyplot as plt

from osgeo import gdal
from shapely.geometry import Point

def read_file(filename):
    with open(filename) as f:
        contents = f.readlines()

    data = []
    for line in contents:
        if line[0:1] == "#":
            continue
        else:
            data.append(line)

    headers = [x.replace("Soil Moisture Percent","smp").replace(" ","_") for x in data[0].split(",")]
    cols = [x.strip("\n").split(",") for x in data[1:]]

    df = pd.DataFrame(cols, columns = headers)
    
    return df

def get_site_lat_lons(site_num, stations_csv):
    df = pd.read_csv(stations_csv)
    site = df[df['site_name'].str.contains(site_no)]
    
    lat, lon = site['lat'].values, site['lon'].values
    return lat[0], lon[0]

csvs = glob.glob("*.csv")
txts = glob.glob("*.txt")
df = pd.read_csv(csvs[0])
lats = df.lat
longs = df.lon

lats = df.lat
longs = df.lon

crs = {"init":"epsg:4326"}
geom = [Point(xy) for xy in zip(longs,lats)]

gdf = gp.GeoDataFrame(df, crs = crs, geometry = geom)

# gdf.to_file("../shape/scan_sites.shp")

# Extract raster values at points:
clays = []
silts = []
sands = []
A = []
B = []

# Read input shapefile with fiona and iterate over each feature
with fiona.open('../shape/scan_sites.shp') as shp:
    for feature in shp:
        station_name = feature['properties']['site_name']
        coords = feature['geometry']['coordinates']
        # Read pixel value at the given coordinates using Rasterio
        # NB: `sample()` returns an iterable of ndarrays.
        with rasterio.open('../rasters/Unified_NA_Soil_Map_Subsoil_Clay_Fraction.tif') as src:
            clay = [v for v in src.sample([coords])][0][0]
        with rasterio.open('../rasters/Unified_NA_Soil_Map_Subsoil_Silt_Fraction.tif') as src:
            silt = [v for v in src.sample([coords])][0][0]
        with rasterio.open('../rasters/Unified_NA_Soil_Map_Subsoil_Sand_Fraction.tif') as src:
            sand = [v for v in src.sample([coords])][0][0]
            
        # Update lists accordingly
        clays.append(clay)
        sands.append(sand)
        silts.append(silt)
        
        # Calculate constants
        a = np.exp(-4.396 - (0.0715*(clay)) - (4.880 * 1e-4*(sand**2)) - (4.285 * 1e-5*(sand**2)*(clay)))
        A.append(a)
        
        b = ((-3.140 - (0.00222 * (clay)**2))) - (3.484 * 1e-5 * ((sand**2)  * (clay)))
        B.append(b)
        

df['clay_fract'] = clays
df['silt_fract'] = silts
df['sand_fract'] = sands
df['a'] = A
df['b'] = B


# df.to_csv("../data/sites_textures.csv")

gdf['clay_fract'] = clays
gdf['silt_fract'] = silts
gdf['sand_fract'] = sands
gdf['a'] = A
gdf['b'] = B

# gdf.to_file("../shape/scan_sites.shp")

gdf.head()
